third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/line-height-expected.txt

This is a testharness.js-based test.
PASS Can set 'line-height' to CSS-wide keywords
PASS Can set 'line-height' to var() references
PASS Can set 'line-height' to the 'normal' keyword
FAIL Can set 'line-height' to a length assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
FAIL Can set 'line-height' to a number assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
FAIL Can set 'line-height' to a percent assert_class_string: relative lengths must compute to a CSSUnitValue expected "[object CSSUnitValue]" but got "[object CSSKeywordValue]"
PASS Setting 'line-height' to a time throws TypeError
PASS Setting 'line-height' to an angle throws TypeError
PASS Setting 'line-height' to a flexible length throws TypeError
PASS Setting 'line-height' to a position throws TypeError
PASS Setting 'line-height' to a URL throws TypeError
PASS Setting 'line-height' to a transform throws TypeError
Harness: the test ran to completion.
